app/requirementsProbe.txt
#fastapi==0.116.1
uvicorn==0.35.0
openai==1.97.1
python-dotenv==1.1.1
python-multipart==0.0.20
streamlit==1.47.0
pydantic==2.11.7
pydantic_core==2.33.2
pandas==2.3.1
python-docx==1.2.0
PyPDF2==3.0.1
lxml==6.0.0
requests==2.32.4
altair==5.5.0
watchdog==6.0.0
tqdm==4.67.1
